+++ bin/maintenance/compile-python-deps.py
@@ -6,11 +6,19 @@
 # modify it under the terms of the MIT License; see the
 # LICENSE file for more details.
 
+import difflib
 import itertools
 import subprocess
 import sys
+from pathlib import Path
 
 import click
+import tomlkit
+from hatch_requirements_txt import load_requirements_files
+from packaging.requirements import Requirement
+from pygments import highlight
+from pygments.formatters.terminal256 import Terminal256Formatter
+from pygments.lexers.diff import DiffLexer
 
 
 COMMANDS = {
@@ -18,6 +26,43 @@ COMMANDS = {
     'dev': ('uv', 'pip', 'compile', 'requirements.dev.in', '-o', 'requirements.dev.txt'),
     'docs': ('uv', 'pip', 'compile', 'docs/requirements.in', '-o', 'docs/requirements.txt'),
 }
+
+
+def _show_diff(old, new, filename):
+    diff = difflib.unified_diff(old.splitlines(), new.splitlines(), filename, filename, lineterm='')
+    diff = '\n'.join(diff)
+    print(highlight(diff, DiffLexer(), Terminal256Formatter(style='native')))
+
+
+def _update_pyproject_reqs(data, context):
+    reqs = {x.name: str(x) for x in load_requirements_files(['requirements.txt'])[0]}
+    reqs |= {x.name: str(x) for x in load_requirements_files(['requirements.dev.txt'])[0]}
+    for getter in (
+        lambda: data['build-system']['requires'],
+        lambda: data['tool']['hatch']['build']['targets']['sdist']['hooks']['custom']['dependencies'],
+    ):
+        try:
+            deplist = getter()
+        except KeyError:
+            continue
+        for i, req in enumerate(deplist):
+            req_name = Requirement(req).name
+            try:
+                new_req = reqs[req_name]
+            except KeyError:
+                print(f'[{context}] Requirement {req_name} not pinned in Indico core')
+                continue
+            deplist[i] = tomlkit.string(new_req, literal=True)
+
+
+def _update_pyproject(pyproject_path: Path, context: str):
+    old_content = pyproject_path.read_text()
+    data = tomlkit.parse(old_content)
+    _update_pyproject_reqs(data, context)
+    new_content = tomlkit.dumps(data)
+    if old_content != new_content:
+        _show_diff(old_content, new_content, str(pyproject_path))
+        pyproject_path.write_text(new_content)
 
 
 @click.command()
@@ -49,6 +94,10 @@ def main(upgrade, upgrade_packages):
         proc = subprocess.run([*command, *args], stdout=subprocess.DEVNULL)
         if proc.returncode:
             failed = True
+
+    _update_pyproject(Path('pyproject.toml'), 'Indico')
+    for plugin_pyproject in Path('../plugins').glob('**/pyproject.toml'):
+        _update_pyproject(plugin_pyproject, plugin_pyproject.parent.name)
 
     sys.exit(1 if failed else 0)